Refactor tests: isolate test using errors.Unwrap
Isolate TestWrappedError_IsCompatibleWithErrorsUnwrap which uses
errors.Unwrap (which appeared with Go 1.13) in a separate test source
compiled only for Go 1.13+.
